Fixed some lint warnings
