+++ package.json
@@ -23,8 +23,11 @@
   },
   "homepage": "https://github.com/PIVX-Labs/MyPIVXWallet#readme",
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.47.1",
+    "@typescript-eslint/parser": "^5.47.1",
     "clean-webpack-plugin": "^4.0.0",
     "css-loader": "^6.7.3",
+    "eslint": "^8.31.0",
     "file-loader": "^6.2.0",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.0",
@@ -46,6 +49,7 @@
     "bootstrap-4": "^4.0.0",
     "bs58": "^5.0.0",
     "buffer": "^6.0.3",
+    "create-xpub": "^2.1.0",
     "hdkey": "^2.0.1",
     "jdenticon": "^3.2.0",
     "jquery": "^3.6.3",
